Move redux actions to common place
